PDFBOX-2423: Change type of getCount() as suggested by Maruan
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1634680 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/PDPageTree.java
@@ -280,7 +280,7 @@ public class PDPageTree implements COSObjectable, Iterable<PDPage>
      * Returns the number of leaf nodes (page objects) that are descendants of this root within the
      * page tree.
      */
-    public long getCount() // todo: make this an 'int'? What is the page limit?
+    public int getCount()
     {
         return root.getInt(COSName.COUNT, 0);
     }
